Fixes #116; Version 0.10.4
