+++ README.md
@@ -12,7 +12,7 @@ A Flutter widget for rendering static html tags as Flutter widgets. (Will render
 Add the following to your `pubspec.yaml` file:
 
     dependencies:
-      flutter_html: ^0.10.3
+      flutter_html: ^0.10.4
 
 ## Currently Supported HTML Tags:
 `a`, `abbr`, `acronym`, `address`, `article`, `aside`, `b`, `bdi`, `bdo`, `big`, `blockquote`, `body`, `br`, `caption`, `cite`, `code`, `data`, `dd`, `del`, `dfn`, `div`, `dl`, `dt`, `em`, `figcaption`, `figure`, `footer`, `h1`, `h2`, `h3`, `h4`, `h5`, `h6`, `header`, `hr`, `i`, `img`, `ins`, `kbd`, `li`, `main`, `mark`, `nav`, `noscript`, `ol`, `p`, `pre`, `q`, `rp`, `rt`, `ruby`, `s`, `samp`, `section`, `small`, `span`, `strike`, `strong`, `sub`, `sup`, `table`, `tbody`, `td`, `template`, `tfoot`, `th`, `thead`, `time`, `tr`, `tt`, `u`, `ul`, `var`
@@ -90,13 +90,22 @@ Check out the official Flutter WebView package here: https://pub.dartlang.org/pa
         }
       },
       customTextAlign: (dom.Node node) {
-          if (node is dom.Element) {
-            switch (node.localName) {
-              case "p":
-                return TextAlign.justify;
-            }
+        if (node is dom.Element) {
+          switch (node.localName) {
+            case "p":
+              return TextAlign.justify;
           }
-        },
+        }
+      },
+      customTextStyle: (dom.Node node, TextStyle baseStyle) {
+        if (node is dom.Element) {
+          switch (node.localName) {
+            case "p":
+              return baseStyle.merge(TextStyle(height: 2, fontSize: 20));
+          }
+        }
+        return baseStyle;
+      },
     )
 
 ## `useRichText` parameter
