+++ lib/flutter_html.dart
@@ -2,6 +2,7 @@ library flutter_html;
 
 import 'package:flutter/material.dart';
 import 'package:flutter_html/html_parser.dart';
+import 'package:flutter_html/rich_text_parser.dart';
 import 'image_properties.dart';
 
 class Html extends StatelessWidget {
