+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: flutter_html
 description: A Flutter widget for rendering static html tags as Flutter widgets. (Will render over 70 different html tags!)
-version: 0.10.3
+version: 0.10.4
 author: Matthew Whitaker <sub6resources@gmail.com>
 homepage: https://github.com/Sub6Resources/flutter_html
